requirements.txt
python-telegram-bot==21.3
google-generativeai>=0.5.0
pdfplumber>=0.10.0
python-docx>=1.1.0
python-dotenv>=1.0.0
